CI: update Github actions to v3
There is a deprecation warning in Github's actions about Node.js 12
actions.  The suggestion is to update the relevant actions to a newer
version, so that they will the Node.js 16 environment.

One missing action from this update is the "actions-rs/toolchain".
Unfortunately, it seems there is no update yet for this action.